Implement separate map fragment for vip or guard
The same map fragment was in use by both guardian and vip roles. They
should be different because they differ on behaviour.

* Create a base map fragment
* Extend base fragment
