+++ app/src/main/java/es/udc/apm/familycare/VipActivity.java
@@ -19,6 +19,7 @@ import android.widget.Toast;
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import es.udc.apm.familycare.interfaces.RouterActivity;
+import es.udc.apm.familycare.maps.VipMapFragment;
 import es.udc.apm.familycare.members.DetailMemberFragment;
 import es.udc.apm.familycare.members.MemberListFragment;
 import es.udc.apm.familycare.utils.Constants;
@@ -43,7 +44,7 @@ public class VipActivity extends AppCompatActivity implements RouterActivity {
                     case R.id.navigation_map_vip:
                         // Don't reload map, heavy operation
                         if(navigation.getSelectedItemId() != item.getItemId()) {
-                            navigate(new CustomMapFragment(), null);
+                            navigate(new VipMapFragment(), null);
                         }
                         return true;
                     case R.id.navigation_members:
@@ -149,7 +150,7 @@ public class VipActivity extends AppCompatActivity implements RouterActivity {
             this.currentScreen = SCREEN_DETAIL;
         } else if (fragment instanceof MemberListFragment) {
             this.currentScreen = SCREEN_MEMBERS;
-        } else if (fragment instanceof CustomMapFragment){
+        } else if (fragment instanceof VipMapFragment){
             this.currentScreen = SCREEN_MAP;
         } else if (fragment instanceof LinkFragment){
             this.currentScreen = SCREEN_LINK;
