+++ app/src/main/java/es/udc/apm/familycare/maps/CustomMapFragment.java
@@ -1,4 +1,4 @@
-package es.udc.apm.familycare;
+package es.udc.apm.familycare.maps;
 
 
 import android.annotation.SuppressLint;
@@ -24,7 +24,8 @@ import android.widget.Toast;
 import com.google.android.gms.common.GooglePlayServicesNotAvailableException;
 import com.google.android.gms.common.GooglePlayServicesRepairableException;
 import com.google.android.gms.common.api.Status;
-import com.google.android.gms.location.FusedLocationProviderClient;import com.google.android.gms.location.LocationServices;
+import com.google.android.gms.location.FusedLocationProviderClient;
+import com.google.android.gms.location.LocationServices;
 import com.google.android.gms.location.places.Place;
 import com.google.android.gms.location.places.ui.PlaceAutocomplete;
 import com.google.android.gms.maps.CameraUpdateFactory;
@@ -47,33 +48,27 @@ import butterknife.ButterKnife;
 import butterknife.OnClick;
 import butterknife.OnFocusChange;
 import butterknife.OnTextChanged;
-import es.udc.apm.familycare.maps.GeofenceRemoveService;
-import es.udc.apm.familycare.maps.GeofenceUpdateService;
-import es.udc.apm.familycare.maps.GeolocationService;
+import es.udc.apm.familycare.R;
 
 
-public class CustomMapFragment extends Fragment implements OnMapReadyCallback {
+public abstract class CustomMapFragment extends Fragment implements OnMapReadyCallback {
 
     private static final String TAG = "CustomMapFragment";
     private static final int PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION = 1;
-    private static final int MAX_CIRCLE_RADIUS = 500;
-    private static final int DEFAULT_CIRCLE_RADIUS = 100;
     private static final int PLACE_AUTOCOMPLETE_REQUEST_CODE = 1;
     private static final float DEFAULT_ZOOM = 15f;
     private static final float SEARCH_ZOOM = 18f;
 
+    private FloatingActionButton mAcceptButton;
+    private FloatingActionButton mDeleteButton;
+    private SeekBar mSeekBar;
     private Boolean mLocationPermissionGranted = false;
-    private HashMap<LatLng, Circle> mCircleHashMap = new HashMap<>();
     private GoogleMap mMap;
     private MapView mMapView;
     private Location mLastKnownLocation = null;
     private CameraPosition mCameraPosition = null;
-    private FloatingActionButton mAcceptButton;
-    private FloatingActionButton mDeleteButton;
-    private SeekBar mSeekBar;
-    private FusedLocationProviderClient mFusedLocationClient;
-    private Marker lastMarker;
 
+    private FusedLocationProviderClient mFusedLocationClient;
 
     @BindView(R.id.et_map_search)
     EditText etSearch;
@@ -82,112 +77,10 @@ public class CustomMapFragment extends Fragment implements OnMapReadyCallback {
     @BindView(R.id.btn_map_clear)
     ImageView btnClear;
 
-    private void showButtonLayer() {
-        mAcceptButton.setVisibility(View.VISIBLE);
-        mDeleteButton.setVisibility(View.VISIBLE);
-        mSeekBar.setVisibility(View.VISIBLE);
-    }
-
-    private void hideButtonLayer() {
-        mAcceptButton.setVisibility(View.GONE);
-        mDeleteButton.setVisibility(View.GONE);
-        mSeekBar.setVisibility(View.GONE);
-        mSeekBar.setOnSeekBarChangeListener(null);
-    }
-
-    private void setMarker(LatLng point) {
-        if (lastMarker != null) {
-            mCircleHashMap.get(lastMarker.getPosition()).remove();
-            lastMarker.remove();
-        }
-
-        lastMarker = mMap.addMarker(new MarkerOptions()
-                .position(point)
-                .icon(BitmapDescriptorFactory
-                        .defaultMarker(BitmapDescriptorFactory.HUE_BLUE)));
-
-        Circle circle = mMap.addCircle(new CircleOptions()
-                .center(point)
-                .radius(DEFAULT_CIRCLE_RADIUS)
-                .strokeColor(Color.BLUE));
-        showButtonLayer();
-
-        mCircleHashMap.put(point, circle);
-
-        mAcceptButton.setOnClickListener(v -> {
-            lastMarker = null;
-            hideButtonLayer();
-            addGeofenceFromCircle(circle);
-        });
-
-        mDeleteButton.setOnClickListener(v -> {
-            circle.remove();
-            hideButtonLayer();
-            lastMarker.remove();
-        });
-
-        mSeekBar.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
-            @Override
-            public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-                circle.setRadius(progress);
-            }
-
-            @Override
-            public void onStartTrackingTouch(SeekBar seekBar) {
-            }
-
-            @Override
-            public void onStopTrackingTouch(SeekBar seekBar) {
-
-            }
-        });
-        mSeekBar.setProgress((int) circle.getRadius());
-
-    }
-
-    private boolean modifyMarker(Marker marker) {
-        if (lastMarker != null && !lastMarker.equals(marker)) {
-            mCircleHashMap.get(lastMarker.getPosition()).remove();
-            lastMarker.remove();
-        }
-
-        showButtonLayer();
-        Circle c = mCircleHashMap.get(marker.getPosition());
-        mDeleteButton.setOnClickListener(v -> {
-            removeGeofenceFromCircle(c);
-            c.remove();
-            hideButtonLayer();
-            marker.remove();
-        });
-
-        mAcceptButton.setOnClickListener(v -> hideButtonLayer());
-
-        mSeekBar.setOnSeekBarChangeListener(new SeekBar.OnSeekBarChangeListener() {
-            @Override
-            public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-                c.setRadius(progress);
-            }
-
-            @Override
-            public void onStartTrackingTouch(SeekBar seekBar) {
-            }
-
-            @Override
-            public void onStopTrackingTouch(SeekBar seekBar) {
-                addGeofenceFromCircle(c);
-            }
-        });
-        mSeekBar.setProgress((int) c.getRadius());
-        return true;
-    }
 
     @Override
     public void onMapReady(GoogleMap googleMap) {
         mMap = googleMap;
-
-        mMap.setOnMapClickListener(this::setMarker);
-
-        mMap.setOnMarkerClickListener(this::modifyMarker);
 
         updateLocationUI();
 
@@ -288,29 +181,31 @@ public class CustomMapFragment extends Fragment implements OnMapReadyCallback {
     }
 
 
+    private void hideButtonLayer() {
+        mAcceptButton.setVisibility(View.GONE);
+        mDeleteButton.setVisibility(View.GONE);
+        mSeekBar.setVisibility(View.GONE);
+        mSeekBar.setOnSeekBarChangeListener(null);
+    }
 
     public CustomMapFragment() {
     }
 
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+    public View inflateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
 
         mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this.getActivity());
 
         View rootView = inflater.inflate(R.layout.fragment_map, container, false);
+
         mAcceptButton = rootView.findViewById(R.id.button_accept);
         mDeleteButton = rootView.findViewById(R.id.button_delete);
         mSeekBar = rootView.findViewById(R.id.seekBar);
-        mSeekBar.setMax(MAX_CIRCLE_RADIUS);
-
-        mFusedLocationClient = LocationServices.getFusedLocationProviderClient(this.getActivity());
+        hideButtonLayer();
 
         this.getActivity().startService(new Intent(this.getActivity(),GeolocationService.class));
 
-        hideButtonLayer();
-
         mMapView = rootView.findViewById(R.id.mapView);
-
+        Log.e("test", "laod map view");
         if (mMapView != null) {
             mMapView.onCreate(savedInstanceState);
             mMapView.getMapAsync(this);
