+++ app/src/main/java/es/udc/apm/familycare/maps/VipMapFragment.java
@@ -1,0 +1,35 @@
+package es.udc.apm.familycare.maps;
+
+
+import android.content.Intent;
+import android.graphics.Color;
+import android.os.Bundle;
+import android.support.design.widget.FloatingActionButton;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.SeekBar;
+
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.model.BitmapDescriptorFactory;
+import com.google.android.gms.maps.model.Circle;
+import com.google.android.gms.maps.model.CircleOptions;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.Marker;
+import com.google.android.gms.maps.model.MarkerOptions;
+
+import java.util.HashMap;
+
+import es.udc.apm.familycare.R;
+
+
+public class VipMapFragment extends CustomMapFragment {
+    private static final String TAG = "VipCustomMapFragment";
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View rootView = super.inflateView(inflater, container, savedInstanceState);
+        return rootView;
+    }
+
+}
